// src/api/__generated__/wallet/model/checkinResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 签到响应
 */
export interface CheckinResponse {
  /** 签到记录ID */
  recordId?: number;
  /** 本周第几天 */
  dayOfWeek?: number;
  /** 获得积分 */
  points?: number;
  /** 签到时间 */
  checkinTime?: string;
}