// src/api/__generated__/wallet/model/checkinWeeklyResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { DayCheckinVO } from './dayCheckinVO';

/**
 * 本周签到信息
 */
export interface CheckinWeeklyResponse {
  /** 本周已签到天数 */
  checkinDays?: number;
  /** 本周总积分 */
  totalPoints?: number;
  /** 签到记录列表 */
  checkinRecords?: DayCheckinVO[];
}